sema_up: restore interrupts before yielding to a higher-priority thread
If invoking sema_up results in the unblocking of a waiting thread that
has a higher priority than the currently running one, then we must yield
to it. The previous implementation called thread_yield while interrupts
were still disabled. While that didn't appear to affect any of the
tests, it seems safer to re-enable interrupts as soon as we no longer
need to access the semaphore's internal data structures, but no later.